feat: enable config cors and finish unit tests

+++ src/main/java/com/codechallenge/openpay/restapisopenpay/service/MarvelService.java
@@ -31,7 +31,7 @@ public class MarvelService {
 
     }
 
-    public Result getCharacterById(int id) {
+    public Result getCharacterById(long id) {
         return marvelApiClientService.getCharacterById(id)
                 .getData()
                 .getResults()
